refactor: move runtime-resolve into vite-plugin-runtime
