--- packages/runtime-resolve/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "@opendatacapture/runtime-resolve",
-  "type": "module",
-  "version": "0.0.0",
-  "license": "Apache-2.0",
-  "exports": {
-    ".": "./src/index.js"
-  },
-  "scripts": {
-    "format": "prettier --write src",
-    "lint": "eslint --fix src"
-  },
-  "peerDependencies": {
-    "@opendatacapture/runtime-v1": "workspace:*"
-  }
-}
