+++ packages/vite-plugin-runtime/src/__tests__/index.test.js
@@ -2,11 +2,11 @@ import fs from 'fs/promises';
 import path from 'path';
 
 import { deepFreeze, range } from '@douglasneuroinformatics/libjs';
-import * as runtimeResolve from '@opendatacapture/runtime-resolve';
 import { afterEach, beforeEach, describe, expect, it } from 'vitest';
 import { vi } from 'vitest';
 
 import { runtime } from '../index.js';
+import * as runtimeResolve from '../resolve.js';
 
 /** @type {runtimeResolve.RuntimeVersionInfo} */
 const runtimeVersionInfoStub = deepFreeze(
